halscope: Replace File/Open and File/Save dialog window
This replaces the open/save dialog which is now called GtkFileChooser.

For the save dialog, it will now prompt if the user tries to overwrite
an existing file. I have also replaced the function "halscope_suffix"
with a built in file filter. This was also added for the "Save Data
File" dialog.