halscope: Change 'Select Channel Source' dialog
To select a new source, a row has to be double-clicked or selected and
pressed enter on the keyboard. This is a change, since previously it was
enough with a single click.
It is easier to search for the correct item, the functionality is also
identical to that in halmeter.

We stopped using the dialog_generic_t struct. Most of the functionality we
used is built in GTK now, so we're using that instead.